scripts/requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
torch>=2.0.0
pyyaml>=6.0
python-frontmatter>=1.0.0
markdown>=3.4.3
beautifulsoup4>=4.12.0
tqdm>=4.65.0
